Module type Make.BOX

The following interface are common elements of multiple modules below. They are used to minimize repetition.

type t

With BOXes, you can put a request right at the front

val put_request : t -> 'a Request.t -> unit

val put_request_and_wait : t -> 'a Request.t -> ('a, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t