Module Tezos_workers.Worker

Worker group maker

type worker_name = {

base : string;

name : string;

}

An error returned when trying to communicate with a worker that has been closed.

type Tezos_base__TzPervasives.Error_monad.error += Closed of worker_name

module type T = sig ... end

Functor to build a group of workers. At that point, all the types are fixed and introspectable, but the actual parameters and event handlers can be tweaked for each individual worker.

module Make : functor (Name : Tezos_base.Worker_intf.NAME) -> functor (Event : Tezos_base.Worker_intf.EVENT) -> functor (Request : Tezos_base.Worker_intf.REQUEST) -> functor (Types : Tezos_base.Worker_intf.TYPES) -> functor (Logger : Tezos_base.Worker_intf.LOGGER with module Event = Event and type Request.view = Request.view) -> T with module Name = Name and module Event = Event and module Request = Request and module Types = Types